Reorganize queries.py (#661)
* Move get_tracks_including_unlisted

* Move get_stems_of

* Move get_feed

* Move get_repost_feed_for_user

* Move get_follow_intersection_users

* Move get_track_repost_intersection_users

* Move get_playlist_repost_intersection_users

* Move get_followers_for_user

* Move get_followees_for_user

* Move get_reposters_for_track

* Move get_reposters_for_playlist

* Move get_savers_for_track

* Move get_savers_for_playlist

* Move get_saves

* Move get_users_account

* Move get_max_id

* Move get_top_playlists

* Move get_top_followee_saves and get_top_followee_windowed

* Move get_top_genre_users and fix to_dict for request.args

* Move get_remixes_of

* Move get_remix_track_parents

* Move get_previously_unlisted_tracks

* Move get_previously_private_playlists

* Clean up exceptions

* Remove quotes in if statement

* Fix a bunch of lint errors
