+++ discovery-provider/pylintrc
@@ -38,7 +38,7 @@ load-plugins=
 # --enable=similarities". If you want to run only the classes checker, but have
 # no Warning level messages displayed, use"--disable=all --enable=classes
 # --disable=W"
-disable=C0111,C0121,R0201,R0912,R0903,R0904,W0142,W0223,W0232,C0103,W0511,R0914,R0921,R0922,R0913,R0801,E1101,E0401,W0613,W0212,logging-fstring-interpolation
+disable=C0111,C0121,C0123,R0201,R0912,R0903,R0904,R1703,W0142,W0223,W0232,C0103,W0511,R0914,R0921,R0922,R0913,R0801,R1718,E1101,E0401,W0613,W0212,W0622,W0703,logging-fstring-interpolation
 
 
 [REPORTS]
